refactor(frontend): move store userSettingsNetworks
