+++ src/frontend/src/lib/derived/user-profile.derived.ts
@@ -1,4 +1,4 @@
-import type { Settings, UserProfile } from '$declarations/backend/backend.did';
+import type { NetworksSettings, Settings, UserProfile } from '$declarations/backend/backend.did';
 import { userProfileStore } from '$lib/stores/user-profile.store';
 import { fromNullishNullable, nonNullish } from '@dfinity/utils';
 import { derived, type Readable } from 'svelte/store';
@@ -20,4 +20,9 @@ export const userProfileVersion: Readable<bigint | undefined> = derived(
 export const userSettings: Readable<Settings | undefined> = derived(
 	[userProfile],
 	([$userProfile]) => fromNullishNullable($userProfile?.settings)
+);
+
+export const userSettingsNetworks: Readable<NetworksSettings | undefined> = derived(
+	[userSettings],
+	([$userSettings]) => $userSettings?.networks
 );
