+++ src/frontend/src/tests/lib/derived/user-networks.derived.spec.ts
@@ -1,5 +1,5 @@
 import { SUPPORTED_MAINNET_NETWORKS_IDS } from '$env/networks/networks.env';
-import { userNetworks, userSettingsNetworks } from '$lib/derived/user-networks.derived';
+import { userNetworks } from '$lib/derived/user-networks.derived';
 import { userProfileStore } from '$lib/stores/user-profile.store';
 import type { UserNetworks } from '$lib/types/user-networks';
 import { mockUserNetworks } from '$tests/mocks/user-networks.mock';
@@ -13,18 +13,6 @@ import { get } from 'svelte/store';
 
 describe('user-networks.derived', () => {
 	const certified = true;
-
-	describe('userSettingsNetworks', () => {
-		it('should return undefined when user profile is not set', () => {
-			userProfileStore.reset();
-			expect(get(userSettingsNetworks)).toBeUndefined();
-		});
-
-		it('should return user profile if it is not nullish', () => {
-			userProfileStore.set({ certified, profile: mockUserProfile });
-			expect(get(userSettingsNetworks)).toEqual(mockNetworksSettings);
-		});
-	});
 
 	describe('userNetworks', () => {
 		const expectedMainnets: UserNetworks = SUPPORTED_MAINNET_NETWORKS_IDS.reduce<UserNetworks>(
